Create scatterplot which shows how detailed/precise the trial is

scatterplot, which shows the distribution of the trial enrollment (logarithm, base 10), and the number of total outcomes (sum of primary, secondary and other outomes) used to measure the results of the trial. The data is also clustered by the sponsor type; in the app the dataset is filtered by the query, but the query acts on the table studies. This table also contains information about enrollment and sposor_type in the columns enrollment and source_class respectively However, there is no information about the number of outcomes specified by each trial. This information is stored in calculated_tables table, hence we left join this table to add this information and filter accordingly. Note, for some studies the information about enrollment or sponsor type is missing. Thus, such observations will not be shown on the scatterplot, but the number of such observations with missing data will be shown in the title of the plot

plot_detailedness(studies_filt, calculated_values)

Arguments

studies_filt

table/tibble/data.frame, input studies table filtered by the query nxd where n is the number of filtered studies and d is the number of features

calculated_values

duckDB table, input calculated_values table which is defined at the very beginning of this file. It is enough to give as an input calculated_values table with the columns nct_id, number_of_primary_outcomes_to_measure, number_of_secondary_outcomes_to_measure and number_of_other_outcomes_to_measure only